# Filter geometries by bbox during decoding