This optimizes geometry decoding. Currently in `tile_datasource` and `tile_datasource_pbf` we:

1) decode all geometries
2) then call `mapnik::geometry::envelope(geom)`
3) then throw out any geometries that don't intersect with the query box.

This change makes this more efficient:

- each geometry part is thrown out on the fly during decoding
- this allows us to throw out parts of multigeometries that don't occur within the query bbox (potentially speeding up rendering which will need to do less work)
- this also allows us to avoid the overhead of calling `mapnik::geometry::envelope(geom)`. (which I've seen take 5-10% of the time in profiling the `vtile-decode` benchmark)

Profiling shows some overhead to calling `bbox.intersect` per vertex for lines and polygons, so the win is slightly less than for multipoints. The `bbox.intersect` overhead however is more like 2-3% of time (vs the 5-10% of time for `mapnik::geometry::envelope(geom)`).

So, overall we see improvements from this pull like:

Before:

```
$ .//build/Release/vtile-decode bench/enf.t5yd5cdi_14_13089_8506.vector.pbf 14 13089 8506 200
z:14 x:13089 y:8506 iterations:200
2822.66ms (cpu 2821.10ms)   | decode as datasource_pbf: bench/enf.t5yd5cdi_14_13089_8506.vector.pbf
2070.90ms (cpu 2070.44ms)   | decode as datasource: bench/enf.t5yd5cdi_14_13089_8506.vector.pbf
processed 6800 features
```

After:

```
$ .//build/Release/vtile-decode bench/enf.t5yd5cdi_14_13089_8506.vector.pbf 14 13089 8506 200
z:14 x:13089 y:8506 iterations:200
2497.01ms (cpu 2493.40ms)   | decode as datasource_pbf: bench/enf.t5yd5cdi_14_13089_8506.vector.pbf
1753.55ms (cpu 1751.10ms)   | decode as datasource: bench/enf.t5yd5cdi_14_13089_8506.vector.pbf
processed 6600 features
```