@springmeyer 2015-08-11
@flippmoke - would love your feedback / quick code review / merge.